fixing some feedback from PR
