+++ frontend/language/src/nb.json
@@ -634,19 +634,19 @@
   "preview.subheader.showas": "Vis som",
   "preview.view_size_desktop": "Desktop",
   "preview.view_size_mobile": "Mobil",
-  "process_editor.configuration_panel_confirmation_task": "Confirmation task",
-  "process_editor.configuration_panel_data_task": "Data task",
+  "process_editor.configuration_panel_confirmation_task": "Task: Bekreftelse",
+  "process_editor.configuration_panel_data_task": "Task: Utfylling",
   "process_editor.configuration_panel_element_not_supported": "Vi støtter foreløpig kun å se på Tasks. Velg en Task for å se detaljer.",
-  "process_editor.configuration_panel_feedback_task": "Feedback task",
-  "process_editor.configuration_panel_header_help_text_confirmation": "Confirmation task brukes for å vise sluttbruker en oppsummering som de kan bekrefte.",
-  "process_editor.configuration_panel_header_help_text_data": "Data task brukes dersom det skal samles inn eller vises frem data.",
-  "process_editor.configuration_panel_header_help_text_feedback": "Feedback task brukes når sluttbruker må vente på tilbakemelding fra tjenesteeier før de kan gå videre i, eller avslutte, prosessen.",
-  "process_editor.configuration_panel_header_help_text_signing": "Signing task brukes når sluttbruker skal bekrefte med signatur.",
+  "process_editor.configuration_panel_feedback_task": "Task: Tilbakemelding",
+  "process_editor.configuration_panel_header_help_text_confirmation": "Bekreftelse-task (confirmation) brukes for å vise sluttbruker en oppsummering som de kan bekrefte.",
+  "process_editor.configuration_panel_header_help_text_data": "Utfylling-task (data) brukes dersom det skal samles inn eller vises frem data.",
+  "process_editor.configuration_panel_header_help_text_feedback": "Tilbakemelding-task (feedback) brukes når sluttbruker må vente på tilbakemelding fra tjenesteeier før de kan gå videre i, eller avslutte, prosessen.",
+  "process_editor.configuration_panel_header_help_text_signing": "Signerings-task (signing) brukes når sluttbruker skal bekrefte med signatur.",
   "process_editor.configuration_panel_header_help_text_title": "Informasjon om valgt task",
   "process_editor.configuration_panel_id_label": "ID:",
   "process_editor.configuration_panel_name_label": "Navn: ",
   "process_editor.configuration_panel_no_task": "Velg et element i prosessen for å se detaljer.",
-  "process_editor.configuration_panel_signing_task": "Signing task",
+  "process_editor.configuration_panel_signing_task": "Task: Signering",
   "process_editor.edit_mode": "Redigeringsmodus",
   "process_editor.fetch_bpmn_error_message": "Kunne ikke laste inn BPMN for denne appen. Prøv igjen senere.",
   "process_editor.fetch_bpmn_error_title": "En feil oppstod ved innlasting av BPMN",
