+++ frontend/packages/process-editor/src/components/ConfigPanel/ConfigPanel.module.css
@@ -7,28 +7,7 @@
   border-left: 1px solid var(--fds-semantic-border-neutral-subtle);
 }
 
-.configPanelNotSelectedText {
+.configPanelParagraph {
   margin-inline: var(--margin-inline);
   margin-top: 1rem;
-}
-
-.headerWrapper {
-  display: flex;
-  align-items: center;
-  margin-inline: var(--margin-inline);
-  margin-top: 1rem;
-  justify-content: space-between;
-}
-
-.headerTextAndIconWrapper {
-  display: flex;
-  align-items: center;
-  gap: 1rem;
-}
-
-.configDetailsRowWrapper {
-  display: flex;
-  flex-direction: column;
-  margin-inline: var(--margin-inline);
-  gap: 1rem;
 }
