+++ frontend/packages/process-editor/src/components/ConfigPanel/ConfigSection/ConfigDetailsRow/ConfigDetailsRow.test.tsx
