+++ frontend/packages/process-editor/src/components/ConfigPanel/ConfigSection/ConfigIcon/ConfigIcon.module.css
