+++ frontend/packages/process-editor/src/components/ConfigPanel/ConfigSection/ConfigIcon/index.ts
