+++ frontend/packages/process-editor/src/components/ConfigPanel/ConfigSection/index.ts
@@ -1,0 +1,1 @@
+export { ConfigSection } from './ConfigSection';
